Divide Controller into Three Controllers And updated some Controller methods
